+++ src/buttons.py
@@ -125,7 +125,9 @@ class Button_selection(Button):
         Parameters
         ----------
         app : Manager(DataGest)
-            Class to update its attribute.
+            The main application instance to update its attributes.
+            Used attribute:
+                pointed by self.target.
 
         """
         if np.any(self.is_mouse_on):
@@ -179,21 +181,22 @@ class Button_app_actions(Button):
     """
     def __init__(self, x_start, x_stop, y_start, y_stop, text, font, lin_w,
                  target:str,
-                 bt_color:list):
+                 bt_color:list) -> None:
 
         super().__init__(x_start, x_stop, y_start, y_stop, text, font, lin_w)
 
         self.target = target
         self.bt_color = bt_color
 
-    def actions(self, app):
+    def actions(self, app) -> None:
         """
         Calls the method specified by 'target' on the app instance.
 
         Parameters
         ----------
         app : Manager(DataGest)
-            Class to update its attribute.
+            The main application instance to update its attributes.
+                pointed by self.target.
 
         """
         if np.any(self.is_mouse_on):
@@ -248,7 +251,7 @@ class Button_keyboard(Button):
     """
     def __init__(self, x_start, x_stop, y_start, y_stop, text, font, lin_w,
                  target:str,
-                 bounds:list):
+                 bounds:list) -> None:
 
         super().__init__(x_start, x_stop, y_start, y_stop, text, font, lin_w)
 
@@ -276,9 +279,11 @@ class Button_keyboard(Button):
 
         Parameters
         ----------
-        app : object
-            The main application instance containing state and error_type
-            attributes.
+        app : Manager(DataGest)
+            The main application instance to update its attributes.
+            Used attribute:
+                state : str, state of the application.
+                error_type : str, type of raised error.
 
         """
         app.state = 'ERROR'
@@ -451,6 +456,214 @@ class Button_keyboard(Button):
                 self.xy_cursor[1]), (self.xy_cursor[0], self.xy_cursor[2]), 3)
 
 
+class Scroll_barr:
+    """
+    Object for the scrolling bar.
+
+    Attributes
+    ----------
+    box : numpy.ndarray
+        X, y up-left corner and width, height value.
+    colors : list
+        The: 1. background color, 2. scoll bar color, 3. contour color RGB.
+    lin_w : int
+        The line width for the button's border.
+    x_st : np.float64
+        X starting value.
+    x_sp : np.float64
+        X stoping value.
+    y_st : np.float64
+        Y starting value.
+    y_sp : np.float64
+        Y stoping value.
+    is_mouse_on : bool
+        If the mouse is on scrolling bar.
+    n_line : int
+        Number of lines in the comparison pannel.
+    show_scroll : bool
+        If the scroll thumb is render or not.
+    scroller : numpy.ndarray
+        (x, y starting rect) and (width & height).
+    delta_h : float
+        The y amount of pixel the thumb bar is moved by wheel iteration.
+    in_box_h : float
+        Total number of pixel where the thumb bar can move.
+    max_travel : float
+        Maximum y value possible for the thumb bar top.
+    max_index : int
+        Maximum index which can be reach by the top of thumb bar.
+        
+    """
+    def __init__(self,
+                 box:np.ndarray,
+                 colors:list,
+                 lin_w:int) -> None:
+
+        self.box = box
+        self.colors = colors
+        self.lin_w = lin_w
+
+        self.x_st = self.box[0]
+        self.x_sp = self.box[0] + self.box[2]
+        self.y_st = self.box[1]
+        self.y_sp = self.box[1] + self.box[3]
+
+        self.is_mouse_on = False
+
+        # initialisation
+        self.n_line = 0
+        self.show_scroll = False
+        self.scroller = [0, 0, 0, 0]
+        self.delta_h = 0
+        self.in_box_h = 0
+        self.max_travel = 0
+        self.max_index = 0
+
+    def test_mouse(self, mouse_pos:tuple) -> None:
+        """
+        Updates the 'is_mouse_on' boolean mask using vectorized comparison.
+
+        Parameters
+        ----------
+        mouse_pos : tuple
+            Current (x, y) coordinates of the mouse.
+
+        """
+        self.is_mouse_on = (mouse_pos[0] >= self.x_st)&(
+                            mouse_pos[1] >= self.y_st)&(
+                            mouse_pos[0] <= self.x_sp)&(
+                            mouse_pos[1] <= self.y_sp)
+
+    def re_init(self) -> None:
+        """
+        Re initialise the scoller bar object.
+        """
+        self.is_mouse_on = False
+
+        # initialisation
+        self.n_line = 0
+        self.show_scroll = False
+        self.scroller = [0, 0, 0, 0]
+        self.delta_h = 0
+        self.in_box_h = 0
+        self.max_travel = 0
+        self.max_index = 0
+
+    def initialise_scroller(self, app) -> None:
+        """
+        Function to initialise the scrollbar. It compute the scroll thumb
+        size and its displacement rate.
+
+        Parameters
+        ----------
+        app : Manager(DataGest)
+            The main application instance to update its attributes.
+            Used attribute:
+                liste1 : list, first list of the authors comparison.
+                mx_tx : int, maximum number of shown comparison lines.
+                text_height : float, height of a comparison line.
+                tex_y : int, starting index of the comparison lines.
+                delta_txy : int, true number of shown comparison lines.
+
+        """
+        self.n_line = len(app.liste1)
+        if self.n_line > 0:
+            self.in_box_h = self.y_sp - self.y_st - 2 * self.lin_w
+            if app.mx_tx < self.n_line:
+                self.show_scroll = True
+
+                self.scroller = [
+                    self.x_st + self.lin_w / 2,
+                    self.y_st + self.lin_w,
+                    self.x_sp - self.x_st - self.lin_w,
+                    max(int((app.mx_tx / self.n_line) * self.in_box_h), 10)]
+
+                self.max_travel = self.in_box_h - self.scroller[3]
+                self.max_index  = self.n_line - app.mx_tx
+                self.delta_h = self.max_travel / self.max_index
+
+        # Calculate the number of lines currently printed in comparison panel
+        if (self.n_line - app.tex_y) >= app.mx_tx:
+            app.delta_txy = app.mx_tx
+        else:
+            app.delta_txy = self.n_line - app.tex_y
+
+    def mouse_wheel(self, event:pygame.event.Event, app) -> None:
+        """
+        Handles mouse wheel events for scrolling through comparison panels and
+        updates the display's vertical scroll positions.
+
+        Parameters
+        ----------
+        app : Manager(DataGest)
+            Class to update its attribute.
+            Used attribute:
+                tex_y : int, starting index of the comparison lines.
+                mx_tx : int, maximum number of shown comparison lines.
+        event : pygame.event.Event
+            The mouse wheel event object. It's expected to have 'y' attribute
+            indicating scroll direction (1 for up, -1 for down).
+
+        """
+        if (event.y != 0)&(self.n_line > app.mx_tx):
+            # Determine scroll direction and update the vertical text
+            # offset indexe (app.tex_y)
+            if event.y == 1:
+                app.tex_y = max(app.tex_y-event.y, 0)
+            else: # == -1
+                app.tex_y = min(app.tex_y-event.y, self.n_line-app.mx_tx)
+
+            # Update the position of the scroller (self.scroller[1]) based on
+            # the new vertical text offset. It scales the scrollbar position
+            # proportionally to the text's scroll position.
+            self.scroller[1] = self.lin_w + self.delta_h * app.tex_y
+
+    def click(self, app) -> None:
+        """
+        Handles the selection logic: ...
+
+        Parameters
+        ----------
+        app : Manager(DataGest)
+            The main application instance to update its attributes.
+            Used attribute:
+                mouse_pos : tuple, mouse position.
+                tex_y : int, starting index of the comparison lines.
+
+        """
+        if self.is_mouse_on and (self.delta_h > 0):
+            top = app.mouse_pos[1] - self.scroller[3] / 2
+            top = max(self.y_st+self.lin_w, min(top,
+                      self.y_sp-self.lin_w-self.scroller[3]))
+
+            if top >= (self.y_sp-self.lin_w-self.scroller[3]):
+                index = self.max_index
+            else:
+                index = int(round(top-self.y_st-self.lin_w
+                                 )/self.max_travel*self.max_index)
+    
+                index = min(index, self.max_index)
+
+            self.scroller[1] = top
+            app.tex_y = index
+
+    def draw(self, window:pygame.surface.Surface) -> None:
+        """
+        Renders the scroll bar.
+
+        Parameters
+        ----------
+        window : pygame.surface.Surface
+            Pygame surface object where buttons are draw.
+
+        """
+        pygame.draw.rect(window, self.colors[0], self.box)
+        if self.show_scroll:
+            pygame.draw.rect(window, self.colors[1], self.scroller)
+
+        pygame.draw.rect(window, self.colors[2], self.box, self.lin_w)
+        
+
 class Inidication:
     """
     class to render color fields.
@@ -463,7 +676,9 @@ class Inidication:
         RGB color.
 
     """
-    def __init__(self, box, color):
+    def __init__(self,
+                 box:list,
+                 color:list) -> None:
         self.box = box
         self.color = color